Práctica 2 completada

+++ P2-old/src/include/MQKPObjectAssignmentOperation.h
@@ -23,7 +23,7 @@ protected:
 	 */
 	int _indexKnapsack;
 	int _indexObj;
-	double _deltaFitness;
+	int _deltaFitness;
 
 public:
